refactor: merge systemd services
Connects-to: #47
Connects-to: #48
Change-type: major
