--- bin/balena-connect-proxy
@@ -1,1 +1,0 @@
-./entry.sh
