--- bin/balena-vpn-api
@@ -1,1 +1,0 @@
-./entry.sh
