+++ bin/start.sh
@@ -8,17 +8,15 @@ cd "${WORK_DIR}"
 # shellcheck disable=1090
 test -f "${WORK_DIR}/config/env" && source "${WORK_DIR}/config/env"
 
-service="$(basename "$0")"
-service="${service/balena-/}"
 command="$(command -v node)"
 args=("-r" "source-map-support/register")
-entrypoint="build/src/${service}/app.js"
+entrypoint="build/src/app.js"
 if [ "${NODE_ENV}" = "development" ]; then
 	tsnode="$(command -v ts-node || true)"
 	if [ -x "${tsnode}" ]; then
 		command="${tsnode}"
 		args=("--files")
-		entrypoint="src/${service}/app.ts"
+		entrypoint="src/app.ts"
 	fi
 fi
