--- config/services/open-balena-vpn-api.service
@@ -1,34 +1,0 @@
-# Copyright (C) 2015 Balena Ltd.
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero General Public License as published
-# by the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Affero General Public License for more details.
-#
-# You should have received a copy of the GNU Affero General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-[Unit]
-Description=open-balena-vpn-api
-Requires=confd.service haproxy.service balena-root-ca.service
-After=confd.service balena-root-ca.service
-StartLimitIntervalSec=0
-
-[Service]
-StandardOutput=journal+console
-StandardError=journal+console
-WorkingDirectory=/usr/src/app
-EnvironmentFile=/usr/src/app/config/env
-ExecStartPre=/bin/bash -c 'mkdir -p /dev/net; if [ ! -c /dev/net/tun ]; then mknod /dev/net/tun c 10 200; fi'
-ExecStartPre=/usr/sbin/iptables-legacy -P FORWARD ${IPTABLES_FORWARD_POLICY}
-ExecStart=/usr/src/app/bin/balena-vpn-api
-Restart=always
-SyslogIdentifier=vpn
-
-[Install]
-WantedBy=basic.target
