+++ src/utils/device.ts
@@ -18,8 +18,8 @@
 import * as Bluebird from 'bluebird';
 import * as _ from 'lodash';
 
-import * as utils from '../utils';
-import { APIError, captureException } from '../utils/errors';
+import { balenaApi } from '.';
+import { APIError, captureException } from './errors';
 
 const authHeader = (auth?: Buffer): { Authorization?: string } => {
 	const headers: { Authorization?: string } = {};
@@ -34,7 +34,7 @@ export interface DeviceInfo {
 	is_connected_to_vpn: boolean;
 }
 
-const getDeviceByUUIDQuery = utils.balenaApi.prepare<{ uuid: string }>({
+const getDeviceByUUIDQuery = balenaApi.prepare<{ uuid: string }>({
 	resource: 'device',
 	options: {
 		$select: ['id', 'is_connected_to_vpn'],
@@ -59,7 +59,7 @@ export const getDeviceByUUID = (
 			throw new APIError(err.message);
 		});
 
-const canAccessDeviceQuery = utils.balenaApi.prepare<{ id: number }>({
+const canAccessDeviceQuery = balenaApi.prepare<{ id: number }>({
 	method: 'POST',
 	resource: 'device',
 	id: { '@': 'id' },
@@ -87,7 +87,7 @@ export const getDeviceVpnHost = (
 	uuid: string,
 	auth?: Buffer,
 ): Bluebird<string> =>
-	utils.balenaApi
+	balenaApi
 		.get({
 			resource: 'service_instance',
 			options: {
