+++ src/utils/index.ts
@@ -15,10 +15,6 @@
 	along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-import { metrics } from '@balena/node-metrics-gatherer';
-import * as cluster from 'cluster';
-import * as express from 'express';
-import * as _ from 'lodash';
 import * as winston from 'winston';
 
 import { PinejsClientRequest } from 'pinejs-client-request';
@@ -33,10 +29,14 @@ export const balenaApi = new PinejsClientRequest(
 );
 export const apiKey = process.env.VPN_SERVICE_API_KEY;
 
-export const getLogger = (service: string, workerId?: string | number) => {
+export const getLogger = (
+	service: string,
+	workerId?: string | number,
+	serviceId?: number,
+) => {
 	let workerLabel = 'master';
-	if (workerId != null) {
-		workerLabel = `${workerId}`;
+	if (workerId != null && serviceId != null) {
+		workerLabel = `${serviceId}.${workerId}`;
 	}
 	const transport = new winston.transports.Console({
 		format: winston.format.combine(
@@ -56,31 +56,15 @@ export const getLogger = (service: string, workerId?: string | number) => {
 	});
 };
 
-export const metricsServer = (handler?: express.RequestHandler) => {
-	const app = express();
-	app.disable('x-powered-by');
-	app.get('/ping', (_req, res) => res.send('OK'));
-	app.get('/cluster_metrics', handler || metrics.aggregateRequestHandler());
-	return app;
-};
-
-export const spawnChildren = (n: number, logger: winston.Logger) => {
-	logger.info(`spawning ${n} worker${n > 1 ? 's' : ''}`);
-	_.times(n, i => {
-		const workerId = i + 1;
-		const restartWorker = (code?: number, signal?: string) => {
-			if (signal != null) {
-				logger.crit(`worker-${workerId} killed with signal ${signal}`);
-			}
-			if (code != null) {
-				logger.crit(`worker-${workerId} exited with code ${code}`);
-			}
-			const env = {
-				...process.env,
-				WORKER_ID: workerId,
-			};
-			cluster.fork(env).on('exit', restartWorker);
-		};
-		restartWorker();
-	});
-};
+import * as clients from './clients';
+export { clients };
+import * as device from './device';
+export { device };
+import * as errors from './errors';
+export { errors };
+export { HAProxy } from './haproxy';
+export { describeMetrics, Metrics } from './metrics';
+export { Netmask } from './netmask';
+export { VpnManager } from './openvpn';
+export { pooledRequest } from './request';
+export { service } from './service';
