--- src/vpn-api/utils/index.ts
@@ -1,24 +1,0 @@
-/*
-	Copyright (C) 2018 Balena Ltd.
-
-	This program is free software: you can redistribute it and/or modify
-	it under the terms of the GNU Affero General Public License as published
-	by the Free Software Foundation, either version 3 of the License, or
-	(at your option) any later version.
-
-	This program is distributed in the hope that it will be useful,
-	but WITHOUT ANY WARRANTY; without even the implied warranty of
-	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-	GNU Affero General Public License for more details.
-
-	You should have received a copy of the GNU Affero General Public License
-	along with this program.  If not, see <https://www.gnu.org/licenses/>.
-*/
-
-import * as clients from './clients';
-export { clients };
-export { HAProxy } from './haproxy';
-export { Netmask } from './netmask';
-export { VpnManager } from './openvpn';
-export { pooledRequest } from './request';
-export { service } from './service';
